refactor: 컴포넌트별 css파일 분리 및 folder-body 영역 컴포넌트로 병합
