+++ src/components/FolderCard.css
@@ -1,0 +1,78 @@
+.folder-card {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-between;
+  gap: 20px;
+  padding: 0;
+  background-color: #f5f5f5;
+  box-shadow: 0px 5px 25px 0px rgba(0, 0, 0, 0.08);
+}
+
+.folder-card .link-card {
+  margin-bottom: 25px;
+  width: calc(33% - 20px);
+  box-sizing: border-box;
+  box-shadow: 0px 5px 25px 0px rgba(0, 0, 0, 0.08);
+  overflow: hidden;
+}
+
+.folder-card .link-card:last-child {
+  margin-bottom: 0;
+}
+
+.folder-card .link-card:hover img {
+  transform: scale(1.3);
+  transition: transform 0.3s ease;
+}
+
+.folder-card .link-card img {
+  width: 100%;
+  height: 250px;
+  flex-shrink: 0;
+}
+
+.folder-card .link-card .info-area {
+  display: flex;
+  width: 100%;
+  padding: 15px 20px;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 10px;
+  border-radius: 0px 0px 15px 15px;
+  background-color: #fff;
+  position: relative;
+  z-index: 1;
+}
+
+.folder-card .link-card .info-area .time-ago {
+  color: #666;
+}
+
+.folder-card .link-card .info-area .title {
+  width: 100%;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+  font-size: 16px;
+  line-height: 24px;
+}
+
+.folder-card .link-card .info-area .date {
+  overflow: hidden;
+  color: #333;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+  font-size: 14px;
+}
+
+@media only screen and (max-width: 1124px) {
+  .folder-card .link-card {
+    width: calc(50% - 20px);
+  }
+}
+
+@media only screen and (max-width: 767px) {
+  .folder-card .link-card {
+    width: 100%;
+  }
+}
