+++ src/components/Footer.js
@@ -1,3 +1,5 @@
+import './Footer.css';
+
 function Footer() {
   return (
     <footer>
