+++ src/styles/index.css
@@ -11,161 +11,6 @@ li {
   align-items: center;
 }
 
-nav {
-  display: flex;
-  justify-content: center;
-  position: sticky;
-  top: 0;
-  width: 100%;
-  background-color: #edf7ff;
-}
-
-.gnb {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-  height: 9.4rem;
-  max-width: 192rem;
-  padding: 0 20rem;
-}
-
-.gnb .profile {
-  display: flex;
-  align-items: center;
-  gap: 6px;
-}
-
-.gnb .profile img {
-  width: 28px;
-  height: 28px;
-}
-
-.gnb .profile span {
-  color: #373740;
-  font-size: 14px;
-  line-height: normal;
-}
-
-.folder-header {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 100%;
-  padding: 20px 0 60px 0;
-  gap: 20px;
-  background-color: #edf7ff;
-}
-
-.folder-header .folder-info {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 12px;
-}
-
-.folder-header .folder-info .img {
-  width: 60px;
-  height: 60px;
-}
-
-.folder-header .folder-info span.name {
-  font-size: 16px;
-  line-height: 24px;
-}
-
-.folder-header p.favorite {
-  font-size: 36px;
-}
-
-.folder-body {
-  margin: 40px 0;
-  padding: 0 20rem;
-}
-
-.folder-body .search-area {
-  display: flex;
-  margin-bottom: 40px;
-  padding: 15px 16px;
-  align-items: center;
-  border-radius: 10px;
-  background: #f5f5f5;
-}
-
-.folder-body .search-area .link-search {
-  margin-left: 10px;
-  width: 100%;
-  border: none;
-  background-color: #f5f5f5;
-}
-
-.folder-card {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-between;
-  gap: 20px;
-  padding: 0;
-  background-color: #f5f5f5;
-  box-shadow: 0px 5px 25px 0px rgba(0, 0, 0, 0.08);
-}
-
-.folder-card .link-card {
-  margin-bottom: 25px;
-  width: calc(33% - 20px);
-  box-sizing: border-box;
-  box-shadow: 0px 5px 25px 0px rgba(0, 0, 0, 0.08);
-  overflow: hidden;
-}
-
-.folder-card .link-card:last-child {
-  margin-bottom: 0;
-}
-
-.folder-card .link-card:hover img {
-  transform: scale(1.3);
-  transition: transform 0.3s ease;
-}
-
-.folder-card .link-card img {
-  width: 100%;
-  height: 250px;
-  flex-shrink: 0;
-}
-
-.folder-card .link-card .info-area {
-  display: flex;
-  width: 100%;
-  padding: 15px 20px;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: 10px;
-  border-radius: 0px 0px 15px 15px;
-  background-color: #fff;
-  position: relative;
-  z-index: 1;
-}
-
-.folder-card .link-card .info-area .time-ago {
-  color: #666;
-}
-
-.folder-card .link-card .info-area .title {
-  width: 100%;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  font-size: 16px;
-  line-height: 24px;
-}
-
-.folder-card .link-card .info-area .date {
-  overflow: hidden;
-  color: #333;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  font-size: 14px;
-}
-
 header {
   display: flex;
   flex-direction: column;
@@ -226,11 +71,6 @@ header {
 }
 
 @media only screen and (max-width: 1199px) {
-  nav .gnb {
-    padding: 0 32px;
-    width: 799px;
-  }
-
   header .hero-header .hero-image {
     width: 650px;
     height: auto;
@@ -241,36 +81,7 @@ header {
   }
 }
 
-@media only screen and (max-width: 1124px) {
-  .folder-body {
-    padding: 0 32px;
-  }
-
-  .folder-card .link-card {
-    width: calc(50% - 20px);
-  }
-}
-
 @media only screen and (max-width: 767px) {
-  nav .gnb {
-    padding: 0 32px;
-    width: 326px;
-  }
-
-  nav .gnb img {
-    width: 88px;
-  }
-
-  nav .gnb a.cta-short {
-    width: 80px;
-    height: auto;
-    padding: 10px 16px;
-  }
-
-  nav .gnb a.cta-short span {
-    font-size: 14px;
-  }
-
   header .hero-header {
     width: 100%;
     padding: 28px 32px 0;
@@ -289,13 +100,6 @@ header {
 
   header .hero-header .hero-image {
     width: 300px;
-  }
-
-  .folder-body {
-    padding: 0 32px;
-  }
-  .folder-card .link-card {
-    width: 100%;
   }
 }
 
@@ -446,70 +250,5 @@ section:nth-of-type(even) {
   section .content-image {
     width: 100%;
     margin-bottom: 16px;
-  }
-}
-
-footer {
-  display: flex;
-  justify-content: center;
-  width: 100%;
-  height: 16rem;
-  padding-top: 3.2rem;
-  background-color: var(--black);
-}
-
-.footer-box {
-  display: flex;
-  justify-content: space-between;
-  width: 100%;
-  max-width: 192rem;
-  height: fit-content;
-  padding: 0 10.4rem;
-}
-
-.copyright {
-  color: #676767;
-  font-family: Arial;
-  font-size: 1.6rem;
-}
-
-.footer-links {
-  display: flex;
-  column-gap: 3rem;
-  padding-right: 1.8rem;
-}
-
-.footer-link {
-  color: #cfcfcf;
-  font-family: Arial;
-  font-size: 1.6rem;
-}
-
-.sns {
-  display: flex;
-  column-gap: 1.2rem;
-  height: 2rem;
-}
-
-@media only screen and (max-width: 767px) {
-  footer {
-    display: block;
-    margin-top: 40px;
-    padding: 0;
-  }
-
-  footer .footer-box {
-    position: relative;
-    padding: 32px 32px 64px 32px;
-    height: 100%;
-  }
-
-  footer .footer-links {
-    gap: 40px;
-  }
-
-  footer .footer-box .copyright {
-    position: absolute;
-    bottom: 30px;
   }
 }
